# openapi.json
{
  "openapi": "3.1.0",
  "info": {"title": "Fal Studio Plugin API", "version": "1.0.0", "description": "A server-side fal.ai connector for image and video generation with reusable identity profiles."},
  "servers": [{"url": "https://YOUR-DOMAIN"}],
  "paths": {
    "/health": {"get": {"operationId": "health", "summary": "Check service health", "responses": {"200": {"description": "OK"}}}},
    "/models": {"get": {"operationId": "listModels", "summary": "List supported fal.ai models", "responses": {"200": {"description": "Models"}}}},
    "/profiles": {
      "get": {"operationId": "listIdentityProfiles", "summary": "List reusable identity profiles", "responses": {"200": {"description": "Profiles"}}},
      "post": {"operationId": "createIdentityProfile", "summary": "Create a reusable identity profile", "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["name","image_urls"], "properties": {"name": {"type": "string"}, "image_urls": {"type": "array", "items": {"type": "string", "format": "uri"}, "maxItems": 9}, "identity_prompt": {"type": "string"}}}}}}, "responses": {"201": {"description": "Created"}}}
    },
    "/generate/image": {"post": {"operationId": "generateImage", "summary": "Generate or edit an image", "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["prompt"], "properties": {"model": {"type": "string", "enum": ["nano-banana-edit","flux-2","flux-2-flash","flux-2-lora"]}, "prompt": {"type": "string"}, "profile_id": {"type": "string"}, "image_urls": {"type": "array", "items": {"type": "string", "format": "uri"}}, "input": {"type": "object", "additionalProperties": true}}}}}}, "responses": {"200": {"description": "Generation result"}}}},
    "/generate/reference-video": {"post": {"operationId": "generateReferenceVideo", "summary": "Generate a Seedance reference-to-video clip", "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["prompt"], "properties": {"model": {"type": "string", "enum": ["seedance-2-reference-fast","seedance-2-reference"]}, "prompt": {"type": "string"}, "profile_id": {"type": "string"}, "image_urls": {"type": "array", "items": {"type": "string", "format": "uri"}}, "video_urls": {"type": "array", "items": {"type": "string", "format": "uri"}}, "audio_urls": {"type": "array", "items": {"type": "string", "format": "uri"}}, "resolution": {"type": "string", "enum": ["480p","720p","1080p"]}, "duration": {"oneOf": [{"type":"string"},{"type":"number"}]}, "aspect_ratio": {"type": "string", "enum": ["auto","21:9","16:9","4:3","1:1","3:4","9:16"]}, "generate_audio": {"type": "boolean"}}}}}}, "responses": {"200": {"description": "Generation result"}}}},
    "/generate/motion-video": {"post": {"operationId": "generateMotionVideo", "summary": "Transfer motion from a video to a reference image using Kling", "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["image_url","video_url"], "properties": {"model": {"type": "string", "enum": ["kling-v3-motion-standard","kling-v3-motion-pro"]}, "prompt": {"type": "string"}, "image_url": {"type": "string", "format": "uri"}, "video_url": {"type": "string", "format": "uri"}, "character_orientation": {"type": "string", "enum": ["video","image"]}, "input": {"type": "object", "additionalProperties": true}}}}}}, "responses": {"200": {"description": "Generation result"}}}}
  }
}